search-vector-api/requirements.txt
# Flask and web framework dependencies
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Cors==4.0.0
Flask-Marshmallow==1.2.1
Flask-Migrate==4.0.1
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
werkzeug==3.1.0

# Database and ORM
SQLAlchemy==2.0.37
SQLAlchemy-Continuum==1.3.15
SQLAlchemy-Utils==0.41.1
psycopg[binary,pool]==3.2.7
pgvector==0.3.6

# Machine Learning and AI
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.6.1
sentence-transformers==2.2.2
transformers==4.48.1
keybert==0.8.5
huggingface-hub==0.25.0

# API and Schema
jsonschema==4.19.2
marshmallow-sqlalchemy==1.0.0

# Utilities
importlib-resources==6.1.1
python-dotenv==1.0.0
requests==2.31.0